requirements.txt
apparmor==2.13.4
appdirs==1.4.4
application-utility==1.3.2
bauh==0.9.6
beautifulsoup4==4.9.1
btrfsutil==1.2.0
CacheControl==0.12.6
caffeine-ng==3.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
colorama==0.4.3
configobj==5.0.6
contextlib2==0.6.0.post1
cryptography==3.0
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
docopt==0.6.2
ewmh==0.1.6
gestures==0.2.2
grub2-theme-preview==1.5.0
gufw==20.4.0
html5lib==1.1
idna==2.10
layoutswitcherlib==0.8.9
LibAppArmor==2.13.4
louis==3.14.0
msgpack==1.0.0
npyscreen==4.10.5
numpy==1.19.0
oauthlib==3.1.0
ordered-set==3.1.1
packaging==20.4
pacman-mirrors==4.16.4
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psutil==5.7.0
pwquality==1.4.2
pyaml==20.4.0
pycairo==1.19.1
pycparser==2.20
PyGObject==3.36.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pysmbc==1.0.20
PySocks==1.7.1
python-dotenv==0.14.0
python-telegram-bot==12.8
python-xlib==0.27
pytoml==0.1.21
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.42
requests==2.24.0
requests-oauthlib==1.3.0
resolvelib==0.4.0
retrying==1.3.3
scipy==1.5.0
setproctitle==1.1.10
sip==4.19.23
six==1.15.0
soupsieve==1.9.6
team==1.0
terminator==1.92
toml==0.10.1
tornado==6.0.4
tweepy==3.8.0
ufw==0.36
urllib3==1.25.9
webencodings==0.5.1
